refactor(core): extract createPluginStateBridge
Pull the 5 plugin-hook-bridging effects + the derivedStateForHook
memo + the two public projections (publicGrabbedBoxes,
publicLabelInstances) out of init() into a focused factory.

The bridge owns:
  - derivedStateForHook  (fires onStateChange)
  - promptMode watcher   (fires onPromptModeChange)
  - selection-box watcher (fires onSelectionBox)
  - drag-box watcher     (fires onDragBox)
  - element-label watcher (fires onElementLabel)
  - publicGrabbedBoxes   (still used by api.getState)
  - publicLabelInstances (still used by api.getState)

The labelHookInputs gathering logic that read all 6 inputs of the
onElementLabel hook in one place was factored into a tiny named
helper inside the bridge.

This required moving the createOverlayVisibility call earlier in
init() (alongside createSelectionSourceSync) so the bridge can read
selectionVisible / dragVisible. Drops the now-unused untrack import
from index.tsx.

index.tsx: 2757 -> 2637 (-120). 49 event-callbacks + api-methods
specs green.
