+++ packages/react-grab/src/core/index.tsx
@@ -9,7 +9,6 @@ import {
   createResource,
   on,
   mapArray,
-  untrack,
 } from "solid-js";
 import { render } from "solid-js/web";
 import { createGrabStore } from "./store.js";
@@ -28,6 +27,7 @@ import { createOverlayEffects } from "./overlay-effects.js";
 import { createEnterBlocker } from "./enter-blocker.js";
 import { createRendererHost } from "./renderer-host.js";
 import { createOverlayVisibility } from "./overlay-visibility.js";
+import { createPluginStateBridge } from "./plugin-state-bridge.js";
 import { CopyFailedError } from "../errors.js";
 import {
   isKeyboardEventTriggeredByInput,
@@ -786,137 +786,37 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       setSelectionSource: actions.setSelectionSource,
     });
 
-    const publicGrabbedBoxes = createMemo(() =>
-      store.grabbedBoxes.map((box) => ({
-        id: box.id,
-        bounds: box.bounds,
-        createdAt: box.createdAt,
-      })),
-    );
-
-    const publicLabelInstances = createMemo(() =>
-      store.labelInstances.map((instance) => ({
-        id: instance.id,
-        status: instance.status,
-        tagName: instance.tagName,
-        componentName: instance.componentName,
-        createdAt: instance.createdAt,
-      })),
-    );
-
-    const derivedStateForHook = createMemo(() => {
-      const active = isActivated();
-      const dragging = isDragging();
-      const copying = isCopying();
-      const inputMode = isPromptMode();
-      const target = targetElement();
-      const drag = dragBounds();
-      const themeEnabled = pluginRegistry.store.theme.enabled;
-      const selectionBoxEnabled = pluginRegistry.store.theme.selectionBox.enabled;
-      const dragBoxEnabled = pluginRegistry.store.theme.dragBox.enabled;
-      const draggingBeyondThreshold = isDraggingBeyondThreshold();
-      const effectiveTarget = effectiveElement();
-      const justCopied = didJustCopy();
-
-      const isSelectionBoxVisible = Boolean(
-        themeEnabled &&
-        selectionBoxEnabled &&
-        active &&
-        !copying &&
-        !justCopied &&
-        !dragging &&
-        effectiveTarget != null,
-      );
-      const isDragBoxVisible = Boolean(
-        themeEnabled && dragBoxEnabled && active && !copying && draggingBeyondThreshold,
-      );
-
-      return {
-        isActive: active,
-        isDragging: dragging,
-        isCopying: copying,
-        isPromptMode: inputMode,
-        isSelectionBoxVisible,
-        isDragBoxVisible,
-        targetElement: target,
-        dragBounds: drag ? { x: drag.x, y: drag.y, width: drag.width, height: drag.height } : null,
-        grabbedBoxes: [...publicGrabbedBoxes()],
-        labelInstances: [...publicLabelInstances()],
-        selectionFilePath: store.selectionFilePath,
-        toolbarState: currentToolbarState(),
-      };
+    const visibility = createOverlayVisibility({
+      grab,
+      phase,
+      elementSelectors,
+      pluginRegistry,
+      isToolbarSelectHovered: () => isToolbarSelectHovered(),
+      isDraggingBeyondThreshold: () => isDraggingBeyondThreshold(),
+      hasDragPreviewBounds: () => dragPreviewBounds().length > 0,
     });
+    const {
+      selectionVisible,
+      selectionTagName,
+      selectionLabelVisible,
+      dragVisible,
+      contextMenuBounds,
+      contextMenuPosition,
+      contextMenuTagName,
+    } = visibility;
 
-    createEffect(
-      on(derivedStateForHook, (state) => {
-        pluginRegistry.hooks.onStateChange(state);
-      }),
-    );
-
-    createEffect(
-      on(
-        () => {
-          const inputMode = isPromptMode();
-          return {
-            inputMode,
-            position: inputMode ? pointer() : untrack(pointer),
-            target: inputMode ? targetElement() : untrack(targetElement),
-          };
-        },
-        ({ inputMode, position, target }) => {
-          pluginRegistry.hooks.onPromptModeChange(inputMode, {
-            x: position.x,
-            y: position.y,
-            targetElement: target,
-          });
-        },
-      ),
-    );
-
-    createEffect(
-      on(
-        () => [selectionVisible(), selectionBounds(), targetElement()] as const,
-        ([visible, bounds, element]) => {
-          pluginRegistry.hooks.onSelectionBox(Boolean(visible), bounds ?? null, element);
-        },
-      ),
-    );
-
-    createEffect(
-      on(
-        () => [dragVisible(), dragBounds()] as const,
-        ([visible, bounds]) => {
-          pluginRegistry.hooks.onDragBox(Boolean(visible), bounds ?? null);
-        },
-      ),
-    );
-
-    createEffect(
-      on(
-        () => {
-          const visible = labelVisible();
-          return [
-            visible,
-            labelVariant(),
-            visible ? cursorPosition() : untrack(cursorPosition),
-            visible ? targetElement() : untrack(targetElement),
-            store.selectionFilePath,
-            store.selectionLineNumber,
-          ] as const;
-        },
-        ([visible, variant, position, element, filePath, lineNumber]) => {
-          pluginRegistry.hooks.onElementLabel(visible, variant, {
-            x: position.x,
-            y: position.y,
-            content: "",
-            element: element ?? undefined,
-            tagName: element ? getTagName(element) || undefined : undefined,
-            filePath: filePath ?? undefined,
-            lineNumber: lineNumber ?? undefined,
-          });
-        },
-      ),
-    );
+    const { publicGrabbedBoxes, publicLabelInstances } = createPluginStateBridge({
+      grab,
+      pluginRegistry,
+      phase,
+      elementSelectors,
+      visibility,
+      dragBounds,
+      selectionBounds,
+      cursorPosition,
+      isDraggingBeyondThreshold,
+      currentToolbarState,
+    });
 
     const cursorOverride = createCursorOverride({ isActivated, isCopying, isPromptMode });
 
@@ -2214,26 +2114,6 @@ export const init = (rawOptions?: Options): ReactGrabAPI => {
       themeHue: () => pluginRegistry.store.theme.hue,
     });
 
-    const visibility = createOverlayVisibility({
-      grab,
-      phase,
-      elementSelectors,
-      pluginRegistry,
-      isToolbarSelectHovered: () => isToolbarSelectHovered(),
-      isDraggingBeyondThreshold: () => isDraggingBeyondThreshold(),
-      hasDragPreviewBounds: () => dragPreviewBounds().length > 0,
-    });
-    const {
-      selectionVisible,
-      selectionTagName,
-      selectionLabelVisible,
-      dragVisible,
-      labelVariant,
-      labelVisible,
-      contextMenuBounds,
-      contextMenuPosition,
-      contextMenuTagName,
-    } = visibility;
 
     const [contextMenuComponentName] = createComponentNameForElement(() =>
       store.frozenElements.length > 1 ? null : store.contextMenuElement,
